+++ RELEASE_NOTES.md
@@ -17,10 +17,12 @@ This version moves service implementations into their related namespaces, to mak
 
 ### 🗑️ Deprecations
 
+* `BaseCalloutActionProvider` has been renamed to `Callouts.BaseActionProvider`.
 * `Keyboard.KeyboardServices` has been renamed to `Keyboard.Services`.
 * `Keyboard.KeyboardState` has been renamed to `Keyboard.State`.
 * `LocalAutocompleteProvider` has been renamed to `Autocomplete.LocalProvider`.
 * `RemoteAutocompleteProvider` has been renamed to `Autocomplete.RemoteProvider`.
+* `StandardCalloutActionProvider` has been renamed to `Callouts.StandardActionProvider`.
 * `StandardKeyboardBehavior` has been renamed to `Keyboard.StandardBehavior`.
 * `StandardKeyboardActionHandler` has been renamed to `KeyboardAction.StandardHandler`.